feat: add Context and Ticket tests

+++ jam-core/src/main/kotlin/io/forge/jam/core/Verdict.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class Verdict(
     @Serializable(with = ByteArrayHexSerializer::class)
     val target: ByteArray,
-    val age: Int,
+    val age: Long,
     val votes: List<Vote>
 ) : Encodable {
     override fun encode(): ByteArray {
